added users validation and fix update avatar errors
