+++ src/constants/index.js
@@ -7,4 +7,3 @@ export const CLOUDINARY = {
 };
 
 export const TEMP_UPLOAD_DIR = path.join(process.cwd(), 'temp');
-//users
