+++ src/controllers/users.js
@@ -19,15 +19,25 @@ export const updateUserAvatarController = async (req, res, next) => {
   const { userId } = req.params;
 
   const photo = req.file;
+  if (!photo) {
+    next(createHttpError(400, 'No file uploaded!'));
+    return;
+  }
+
+  if (Object.keys(req.body).length > 0) {
+    return next(createHttpError(400, 'Only a photo file can be uploaded.'));
+  }
+
   let photoUrl;
   if (photo) {
     photoUrl = await saveFileToCloudinary(photo);
   }
 
-  const result = await updateUser(userId, { ...req.body, photo: photoUrl });
+  const result = await updateUser(userId, { photo: photoUrl });
 
   if (!result) {
     next(createHttpError(404, 'User not found.'));
+    return;
   }
 
   res.json({
