+++ src/routes/users.js
@@ -7,7 +7,7 @@ import {
 import { isValidId } from '../middlewares/isValidId.js';
 import { upload } from '../middlewares/multer.js';
 import { validateBody } from '../middlewares/validateBody.js';
-import { updateUserSchema } from '../validation/auth.js';
+import { updateUserSchema } from '../validation/users.js';
 
 const usersRouter = Router();
 
@@ -15,7 +15,6 @@ usersRouter.get('/currentUser', getCurrentUserController);
 usersRouter.patch(
   '/updateUserPhoto/:userId',
   isValidId,
-  validateBody(updateUserSchema),
   upload.single('photo'),
   updateUserAvatarController,
 );
